+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 8.0.0-nullsafe.2
+  - Allow start page, per_page, number of pages options to pagination helper
+  - Allow page options for listTags
+
 ## 8.0.0-nullsafe.1
   - Update to null safety